+++ src/main/java/nerdhub/cardinal/components/api/BlockComponentProvider.java
@@ -1,5 +1,6 @@
 package nerdhub.cardinal.components.api;
 
+import nerdhub.cardinal.components.api.component.Component;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
 import net.minecraft.world.BlockView;
@@ -14,15 +15,15 @@ public interface BlockComponentProvider {
      *
      * @return whether or not this {@link BlockComponentProvider} can provide the desired component
      */
-    <T> boolean hasComponent(BlockView blockView, BlockPos pos, ComponentType<T> type, @Nullable Direction side);
+    <T extends Component> boolean hasComponent(BlockView blockView, BlockPos pos, ComponentType<T> type, @Nullable Direction side);
 
     /**
      * @return an instance of the requested component, or {@code null}
      */
-    <T> T getComponent(BlockView blockView, BlockPos pos, ComponentType<T> type, @Nullable Direction side);
+    <T extends Component> T getComponent(BlockView blockView, BlockPos pos, ComponentType<T> type, @Nullable Direction side);
 
     /**
      * @return an <strong>immutable</strong> view of the component types exposed by this {@link BlockComponentProvider}
      */
-    Set<ComponentType<?>> getComponentTypes(BlockView blockView, BlockPos pos, @Nullable Direction side);
+    Set<ComponentType<? extends Component>> getComponentTypes(BlockView blockView, BlockPos pos, @Nullable Direction side);
 }