+++ src/main/java/nerdhub/cardinal/components/mixins/accessor/ItemstackComponents.java
@@ -1,10 +1,10 @@
 package nerdhub.cardinal.components.mixins.accessor;
 
 import nerdhub.cardinal.components.api.ComponentType;
-import nerdhub.cardinal.components.api.component.ItemComponent;
+import nerdhub.cardinal.components.api.component.Component;
 
 public interface ItemstackComponents {
 
-    <T extends ItemComponent> void setComponentValue(ComponentType<T> type, ItemComponent obj);
+    <T extends Component> void setComponentValue(ComponentType<T> type, Component obj);
 
 }