+++ changelog.md
@@ -1,4 +1,10 @@
 ------------------------------------------------------
+Version 1.1.0
+------------------------------------------------------
+- updated component system -> merged item and block components
+- added Identifier to component registry
+
+------------------------------------------------------
 Version 1.0.0
 ------------------------------------------------------
 - Initial Release